// src/types/player/playerStat.type.ts
export type PlayerStat = 
  'mana'
  | 'maxMana'
  | 'attackPower'
  | 'attackSpeed'
  | 'critChance'
  | 'critMulti'
  | 'experience'
  | 'goldCoins'
  | 'level'
  | 'unspentSkillPoints'
  | 'goldCoinsMultiplier'
  | 'manaRegenRate'
  | 'currentManaRegenTimer'
  | 'magicDamage'
  | 'spellCooldownReduction'
  | 'xpMultiplier'
  | 'shopRefreshCooldown'
  | 'currentShopRefreshCooldown'
  | 'extraFireDamage'
  | 'extraAirDamage'
  | 'extraWaterDamage'
  | 'extraEarthDamage'
  | 'extraLightDamage'
  | 'extraDarkDamage'
  | 'extraPhysicalDamage'
  | 'increasedSpellDuration'